cast to avoid warning
git-svn-id: svn://svn.savannah.nongnu.org/gap/trunk@2701 d25b4930-3292-e311-b074-0010c6de5da8

+++ libs/DataBasinKit/DBSoap.m
@@ -646,7 +646,7 @@
       record = [records objectAtIndex:0];
       batchSize = [records count];        
       
-      NSLog(@"records size is: %lu", batchSize);
+      NSLog(@"records size is: %lu", (unsigned long)batchSize);
       
       /* let's get the fields from the keys of the first record */
       keys = [NSMutableArray arrayWithArray:[record allKeys]];
